refactor: layout and theme
