+++ config/config.exs
@@ -99,15 +99,15 @@ config :ex_audit,
     WraftDoc.Document.DataTemplate,
     WraftDoc.Document.Instance.Version,
     WraftDoc.Document.Instance,
-    WraftDoc.Document.LayoutAsset,
-    WraftDoc.Document.Layout,
     WraftDoc.Document.OrganisationField,
     WraftDoc.Document.Pipeline.Stage,
     WraftDoc.Document.Pipeline,
-    WraftDoc.Document.Theme,
+    WraftDoc.Layouts.LayoutAsset,
+    WraftDoc.Layouts.Layout,
     WraftDoc.Enterprise.ApprovalSystem,
     WraftDoc.Enterprise.Flow,
     WraftDoc.Enterprise.Flow.State,
+    WraftDoc.Themes.Theme,
     WraftDoc.Enterprise.Vendor
   ],
   primitive_structs: [
