+++ test/wraft_doc/document/document_test.exs
@@ -24,16 +24,15 @@ defmodule WraftDoc.DocumentTest do
   alias WraftDoc.Document.Instance.History
   alias WraftDoc.Document.Instance.Version
   alias WraftDoc.Document.InstanceApprovalSystem
-  alias WraftDoc.Document.Layout
-  alias WraftDoc.Document.LayoutAsset
   alias WraftDoc.Document.Pipeline
   alias WraftDoc.Document.Pipeline.Stage
   alias WraftDoc.Document.Pipeline.TriggerHistory
-  alias WraftDoc.Document.Theme
-  alias WraftDoc.Document.ThemeAsset
+  alias WraftDoc.Layouts.Layout
+  alias WraftDoc.Layouts.LayoutAsset
   alias WraftDoc.Repo
+  alias WraftDoc.Themes.Theme
+  alias WraftDoc.Themes.ThemeAsset
   alias WraftDoc.Validations.Validation
-
   setup :verify_on_exit!
 
   @valid_layout_attrs %{
